requirements.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==3.4.7
Django==3.2.6
django-environ==0.4.5
django-extensions==3.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
factory-boy==3.2.0
Faker==8.11.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
